feat: fix exports in all packages and move mail-bridge to Hono
