+++ .vscode/cspell.json
@@ -14,6 +14,7 @@
     "hanko",
     "Hanko",
     "hono",
+    "MAILBRIDGE",
     "mailfetch",
     "mailserver",
     "mediumint",
