+++ apps/mail-bridge/postal-db/index.ts
@@ -1,17 +1,14 @@
 import { drizzle } from 'drizzle-orm/mysql2';
 import mysql from 'mysql2/promise';
 import * as schema from './schema';
-import { useRuntimeConfig } from '#imports';
-import type { PostalConfig } from '../types';
+import { activePostalServer, env } from '../env';
 
-const postalConfig = useRuntimeConfig().postal as PostalConfig;
-
-const isLocal = postalConfig.localMode;
+const isLocal = env.MAILBRIDGE_POSTAL_LOCAL_MODE;
 export const connection = mysql.createPool({
   uri: isLocal
     ? // we actually don't use the db in local mode, it is set to the local docker db to avoid throwing connection errors
-      process.env.DB_MYSQL_MIGRATION_URL
-    : `${(useRuntimeConfig().postal as any).activeServers.dbConnectionString}/postal`,
+      env.DB_MYSQL_MIGRATION_URL
+    : `${activePostalServer.dbConnectionString}/postal`,
   multipleStatements: true
 });
