--- apps/mail-bridge/routes/health.get.ts
@@ -1,5 +1,0 @@
-import { eventHandler } from '#imports';
-
-export default eventHandler(() => {
-  return { status: "I'm Alive 🏝️" };
-});
