--- apps/mail-bridge/trpc/createContext.ts
@@ -1,17 +1,0 @@
-import type { inferAsyncReturnType } from '@trpc/server';
-import type { H3Event } from 'h3';
-import { db } from '@u22n/database';
-import { useRuntimeConfig } from '#imports';
-import { getHeader } from 'h3';
-
-//  * Creates context for an incoming request
-//  * @link https://trpc.io/docs/context
-
-export const createContext = async (event: H3Event) => {
-  const config = useRuntimeConfig(event);
-  const authToken = getHeader(event, 'Authorization');
-  const isServiceAuthenticated = authToken === config.key;
-  return { auth: isServiceAuthenticated, db, config };
-};
-
-export type CreateContext = inferAsyncReturnType<typeof createContext>;
