+++ apps/mail-bridge/trpc/index.ts
@@ -1,12 +1,10 @@
 export * from '@trpc/server';
 import { router } from './trpc';
-import { createContext } from './createContext';
 import { orgRouter } from './routers/orgRouter';
 import { domainRouter } from './routers/domainRouter';
 import { sendMailRouter } from './routers/sendMailRouter';
 import { smtpRouter } from './routers/smtpRouter';
 
-export const trpcMailBridgeContext = createContext;
 export const trpcMailBridgePostalRouter = router({
   org: orgRouter,
   domains: domainRouter
