+++ apps/mail-bridge/tsconfig.json
@@ -1,5 +1,3 @@
 {
-  "extends": ["./.nitro/types/tsconfig.json", "@u22n/tsconfig"],
-  "include": ["./**/*.ts"],
-  "exclude": ["node_modules"]
+  "extends": "@u22n/tsconfig"
 }
