+++ apps/mail-bridge/utils/realtime.ts
@@ -1,19 +1,18 @@
-import { useRuntimeConfig } from '#imports';
 import { db } from '@u22n/database';
 import { eq, inArray } from '@u22n/database/orm';
 import { convoEntries, convoParticipants, convos } from '@u22n/database/schema';
 import RealtimeServer from '@u22n/realtime/server';
-import type { TypeId } from '@u22n/utils';
 
-export const realtime = new RealtimeServer(
-  useRuntimeConfig().realtime as {
-    host: string;
-    port: string;
-    appId: string;
-    appKey: string;
-    appSecret: string;
-  }
-);
+import type { TypeId } from '@u22n/utils';
+import { env } from '../env';
+
+export const realtime = new RealtimeServer({
+  host: env.REALTIME_HOST,
+  port: env.REALTIME_PORT,
+  appSecret: env.REALTIME_APP_SECRET,
+  appId: env.REALTIME_APP_ID,
+  appKey: env.REALTIME_APP_KEY
+});
 
 export async function sendRealtimeNotification({
   newConvo,
