+++ apps/platform/trpc/routers/authRouter/passwordRouter.ts
@@ -7,8 +7,12 @@ import {
 } from '../../trpc';
 import { eq } from '@u22n/database/orm';
 import { accounts } from '@u22n/database/schema';
-import { nanoIdToken, typeIdGenerator, zodSchemas } from '@u22n/utils';
-import { strongPasswordSchema } from '@u22n/utils/password';
+import {
+  nanoIdToken,
+  typeIdGenerator,
+  zodSchemas,
+  strongPasswordSchema
+} from '@u22n/utils';
 import { TRPCError } from '@trpc/server';
 import { createError, setCookie } from 'h3';
 import { lucia } from '../../../utils/auth';
