+++ apps/storage/tsup.config.ts
@@ -5,6 +5,5 @@ export default defineConfig({
   format: 'esm',
   treeshake: true,
   noExternal: [/@u22n\/.+/], // We currently don't bundle our monorepo packages
-  clean: true,
-  minify: true
+  clean: true
 });
