+++ packages/tiptap/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "description": "",
   "main": "index.js",
+  "type": "module",
   "exports": {
     ".": "./index.ts",
     "./extensions": "./extensions.ts",
